workers/tkr_qulacs/pyproject.toml
[project]
name = "tkr-qulacs"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pytket-qulacs>=0.41.0",
    "tierkreis>=2.0.2",
]
